#54: Incorporate Feedback
- Fix Logger
- Use constant in routing
- Only set UI Schema folder if we generate temporary workspace
- Split DefaultModelServerModule and generic ModelServerModule
- Move provider defaults into separate class (like with multibindings)
